move non returning textfield class
